gh-450: move all `pytest.fixture` into `conftest.py`
This makes them available to all tests. I need `cosmo` for use in #449.

Have also added `scope="session"` so they are only computed once when
testing rather than multiple times.
